Change type of orderCount to Long

+++ backend/sigma/src/main/java/com/klikk/sigma/entity/Variation.java
@@ -63,7 +63,7 @@ public class Variation {
     private LocalDateTime modifiedAt;
 
     @Column(name = "order_count")
-    private int orderCount;
+    private Long orderCount;
 
     @ManyToOne
     @JoinColumn(name = "parent_id", referencedColumnName = "id")
